Address review feedback for auth middleware
- Hash tokens (SHA-256) before using them as cache keys to avoid
  keeping raw bearer tokens in process memory
- Add tokenCache.delete() method to encapsulate cache internals
- Use strings.EqualFold for case-insensitive role name comparison
  (Keystone role names are case-preserving)
- Change authPolicies check from nil to len==0 to correctly handle
  empty-but-non-nil slices
- Validate that auth.policies is non-empty when auth is configured
- Reject policy patterns with empty method or path components
- Move auth type definitions from shim.go to auth.go for cohesion
